requirements.txt
python-dotenv
Flask
openai
requests